Split string functions to their own header/file
This is chipping away at the monolithic `firestorm.h` problem, trying to split
more generic library functions into board-independent files.
