+++ userland/examples/security_app/main.c
@@ -5,6 +5,7 @@
 #include <stdint.h>
 
 #include <tock.h>
+#include <tock_str.h>
 #include <firestorm.h>
 #include <gpio.h>
