+++ userland/libtock/firestorm.h
@@ -13,10 +13,6 @@
 extern "C" {
 #endif
 
-void putstr(const char* str);
-void putnstr(const char* str, size_t len);
-void putnstr_async(const char* str, size_t len, subscribe_cb cb, void* userdata);
-
 /*
  * Sets the callback for timers
  *
